greatly simplified now called HandleFileSelect component
