+++ app/(root)/dashboard/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useFileDetails } from "../../contexts/FileDetailsContext";
-import { HandleFolderSelect } from "@/app/components/dashboard/HandleFolderSelect";
+import { HandleFileSelect } from "@/app/components/dashboard/HandleFileSelect";
 import { VideoList } from "@/app/components/dashboard/VideoList";
 import { useDetectMobileDevice } from "@/app/hooks/useDetectMobileDevice";
 import { Toaster } from "@/app/components/ui/sonner";
@@ -12,7 +12,7 @@ export default function Dashboard() {
 
   return (
     <main>
-      <HandleFolderSelect
+      <HandleFileSelect
         fileDetails={fileDetails}
         setFileDetails={setFileDetails}
       />
